feat: disable alpha write, enable depth testing

+++ shaders/vertex.glsl
@@ -18,8 +18,8 @@ uniform float uMaxSplatSize;
 in vec3 position;       // Quad corner coordinates (-1 to 1) with local offset in .z
 // Base index computed from gl_InstanceID - no attribute needed
 
-out highp vec4 vColor;
-out highp vec2 vPosition;     // Quad corner coordinates for fragment shader
+out mediump vec4 vColor;
+out mediump vec2 vPosition;     // Quad corner coordinates for fragment shader
 
 uniform int splatCount;
